# change % into - in ^(::perm, ::Integer)