@thofma 2018-07-30
Nice speedup.